Refactored to share code across media and content type repositories; set appropriate sort orders when moving properties to the composite type
