+++ src/Umbraco.Core/Persistence/Repositories/ContentTypeBaseRepository.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Data;
-using System.Diagnostics;
 using System.Globalization;
 using System.Linq;
 using System.Text;
@@ -12,7 +11,6 @@ using Umbraco.Core.Logging;
 using Umbraco.Core.Models;
 using Umbraco.Core.Models.EntityBase;
 using Umbraco.Core.Models.Rdbms;
-
 using Umbraco.Core.Persistence.Factories;
 using Umbraco.Core.Persistence.Querying;
 using Umbraco.Core.Persistence.Relators;
@@ -1225,6 +1223,214 @@ WHERE cmsContentType." + aliasColumn + @" LIKE @pattern",
             string test;
             while (aliases.Contains(test = alias + i)) i++;
             return test;
+        }
+
+        /// <summary>
+        /// Extracts a set of properties from a content type into a new composition type
+        /// </summary>
+        /// <param name="contentType"><see cref="IContentTypeComposition"/> to extract composition from</param>
+        /// <param name="compositionContentType"><see cref="IContentTypeComposition"/> to extract composition to</param>
+        /// <param name="propertyAliases">Aliases of properties to move to composition type</param>
+        protected void ExtractContentTypeComposition(IContentTypeComposition contentType, IContentTypeComposition compositionContentType, string[] propertyAliases)
+        {
+            using (var trans = Database.GetTransaction())
+            {
+                // Copy property groups (tabs) for any properties selected to composition type (if they don't already exist) and
+                // move properties to those groups on the composition type
+                var copiedPropertyGroups = new Dictionary<int, int>();
+                foreach (var propertyAlias in propertyAliases)
+                {
+                    var property = contentType.PropertyTypes
+                        .SingleOrDefault(x => x.Alias == propertyAlias);
+                    if (property == null)
+                    {
+                        throw new NullReferenceException(
+                            string.Format(
+                                "Property with alias '{0}' selected to extract to a composition could not be found on the provided content type",
+                                propertyAlias));
+                    }
+
+                    var propertyGroup = contentType.PropertyGroups
+                        .Single(x => x.Id == property.PropertyGroupId.Value);
+
+                    // Copy the property group, if we haven't copied it already from another property or if it doesn't already exist, and
+                    // get the Id of the new property group on the composition
+                    int compositeGroupId;
+                    if (copiedPropertyGroups.ContainsKey(propertyGroup.Id) == false)
+                    {
+                        // We haven't copied this one yet, but could be it already exists if we are extracting to an existing type
+                        var compositePropertyGroup = compositionContentType.PropertyGroups
+                            .SingleOrDefault(x => x.Name == propertyGroup.Name);
+                        if (compositePropertyGroup == null)
+                        {
+                            compositeGroupId = CopyPropertyGroup(propertyGroup.Id, copiedPropertyGroups.Count, compositionContentType.Id);
+                        }
+                        else
+                        {
+                            compositeGroupId = compositePropertyGroup.Id;
+                        }
+
+                        copiedPropertyGroups.Add(propertyGroup.Id, compositeGroupId);
+                    }
+                    else
+                    {
+                        compositeGroupId = copiedPropertyGroups[propertyGroup.Id];
+                    }
+
+                    // Before we move the property type to the composite, we need to check if one already exists with the same alias, which
+                    // if could if we are extracting to an existing type
+                    var existingProperty = compositionContentType.PropertyTypes
+                        .SingleOrDefault(x => x.Alias == propertyAlias);
+                    if (existingProperty == null)
+                    {
+                        // Move the property to the composition type within the composition
+                        MovePropertyType(property.Id, compositeGroupId, compositionContentType.Id);
+                    }
+                    else
+                    {
+                        // We've got a clashing alias, let's check if the data types match - if so we'll combine the properties and
+                        // any associated content
+                        if (existingProperty.DataTypeDefinitionId == property.DataTypeDefinitionId)
+                        {
+                            MoveContentBetweenPropertyTypes(property.Id, existingProperty.Id);
+                            DeletePropertyType(property.Id);
+                        }
+                        else
+                        {
+                            // Aliases are the same but the data types are different, can't combine these so we need to change the alias
+                            var alias = CreateUniqueAlias(propertyAlias,
+                                compositionContentType.PropertyTypes.Select(x => x.Alias).ToArray());
+                            MovePropertyType(property.Id, compositeGroupId, compositionContentType.Id, alias);
+                        }
+                    }
+                }
+
+                // Create the composition relation
+                CreateCompositionRelation(compositionContentType.Id, contentType.Id);
+
+                trans.Complete();
+            }
+        }
+
+        /// <summary>
+        /// Copies a property group to a new content type
+        /// </summary>
+        /// <param name="propertyGroupId">The property group Id</param>
+        /// <param name="sortOrder">The new sort order for the group</param>
+        /// <param name="contentTypeId">The Id of the content type to move to</param>
+        /// <returns>Id of newly created property group</returns>
+        private int CopyPropertyGroup(int propertyGroupId, int sortOrder, int contentTypeId)
+        {
+            var dto = Database.Single<PropertyTypeGroupDto>("WHERE id = @PropertyGroupId", new { PropertyGroupId = propertyGroupId });
+            dto.ContentTypeNodeId = contentTypeId;
+            dto.SortOrder = sortOrder;
+            dto.UniqueId = Guid.NewGuid();
+            Database.Insert(dto);
+            return dto.Id;
+        }
+
+        /// <summary>
+        /// Moves a property to a new content type
+        /// </summary>
+        /// <param name="propertyId">The property Id</param>
+        /// <param name="propertyGroupId">The Id of the property group to move to</param>
+        /// <param name="contentTypeId">The Id of the content type to move to</param>
+        /// <param name="updatedAlias">New alias to use (if provided)</param>
+        private void MovePropertyType(int propertyId, int propertyGroupId, int contentTypeId, string updatedAlias = "")
+        {
+            var maxSortOrderOnGroup = Database.ExecuteScalar<int>("SELECT Max(sortOrder) " +
+                                                                  "FROM cmsPropertyType " +
+                                                                  "WHERE contentTypeId = @ContentTypeId " +
+                                                                  "AND propertyTypeGroupId = @PropertyGroupId",
+                new
+                {
+                    ContentTypeId = contentTypeId,
+                    PropertyGroupId = propertyGroupId
+                });
+            var dto = Database.Single<PropertyTypeDto>("WHERE id = @PropertyId", new { PropertyId = propertyId });
+            dto.ContentTypeId = contentTypeId;
+            dto.PropertyTypeGroupId = propertyGroupId;
+            dto.SortOrder = maxSortOrderOnGroup + 1;
+            if (string.IsNullOrEmpty(updatedAlias) == false)
+            {
+                dto.Alias = updatedAlias;
+            }
+
+            Database.Update(dto);
+        }
+
+        /// <summary>
+        /// Moves all content associated with a property type to another
+        /// </summary>
+        /// <param name="fromPropertyId">The property Id to move content from</param>
+        /// <param name="toPropertyId">The property Id to move content to</param>
+        private void MoveContentBetweenPropertyTypes(int fromPropertyId, int toPropertyId)
+        {
+            Database.Execute("UPDATE cmsPropertyData " +
+                             "SET propertytypeid = @ToPropertyId " +
+                             "WHERE propertytypeid = @FromPropertyId",
+                new
+                {
+                    FromPropertyId = fromPropertyId,
+                    ToPropertyId = toPropertyId,
+                });
+        }
+
+        /// <summary>
+        /// Deletes a property from a new content type
+        /// </summary>
+        /// <param name="propertyId">The property Id</param>
+        private void DeletePropertyType(int propertyId)
+        {
+            Database.Delete<PropertyTypeDto>("WHERE id = @PropertyId", new { PropertyId = propertyId });
+        }
+
+        /// <summary>
+        /// Creates a composition relation between two document types if it doesn't already exist
+        /// </summary>
+        /// <param name="parentId">Parent type</param>
+        /// <param name="childId">Child type</param>
+        private void CreateCompositionRelation(int parentId, int childId)
+        {
+            var dto = Database.SingleOrDefault<ContentType2ContentTypeDto>("WHERE parentContentTypeId = @ParentId AND childContentTypeId = @ChildId",
+                new
+                {
+                    ParentId = parentId,
+                    ChildId = childId
+                });
+            if (dto == null)
+            {
+                Database.Insert(new ContentType2ContentTypeDto { ParentId = parentId, ChildId = childId });
+            }
+        }
+
+        /// <summary>
+        /// Helper to generate a unique alias
+        /// </summary>
+        /// <param name="alias">Alias that has a duplicate in the list of aliases for the content type</param>
+        /// <param name="existingAliases">The list of existing aliases</param>
+        /// <returns>Generated unique alias</returns>
+        private string CreateUniqueAlias(string alias, string[] existingAliases)
+        {
+            var suffix = 2;
+            var uniqueAliasAttempt = alias;
+            while (existingAliases.Contains(uniqueAliasAttempt))
+            {
+                uniqueAliasAttempt = alias + suffix;
+                suffix++;
+            }
+
+            return uniqueAliasAttempt;
+        }
+
+        /// <summary>
+        /// Checks to see if a given content type is used as a composition on other content types
+        /// </summary>
+        /// <param name="contentTypeId">Id of content type</param>
+        /// <returns>True if used as a composition on another type, otherwise false</returns>
+        public bool IsUsedAsComposition(int contentTypeId)
+        {
+            return Database.FirstOrDefault<ContentType2ContentTypeDto>("WHERE parentContentTypeId = @ContentTypeId", new { ContentTypeId = contentTypeId }) != null;
         }
     }
 }
