+++ src/Umbraco.Core/Services/ContentTypeService.cs
@@ -893,7 +893,104 @@ namespace Umbraco.Core.Services
             IContentType extractIntoType = null, string nameOfNewType = null, int userId = 0)
         {
             Mandate.ParameterNotNull(contentType, "contantType");
+            MandateExtractCompositionParametersValid(contentType, propertyAliases, extractIntoType, nameOfNewType);
+
+            IContentType compositionType;
+            using (new WriteLock(Locker))
+            {
+                var uow = UowProvider.GetUnitOfWork();
+                using (var repository = RepositoryFactory.CreateContentTypeRepository(uow))
+                {
+                    if (extractIntoType == null)
+                    {
+                        // Create new composition type
+                        compositionType = new ContentType(contentType.ParentId)
+                        {
+                            Alias = nameOfNewType.ToSafeAlias(true),
+                            Name = nameOfNewType,
+                            Icon = "icon-document",
+                        };
+                        repository.AddOrUpdate(compositionType);
+                        uow.Commit();
+                    }
+                    else
+                    {
+                        compositionType = extractIntoType;
+                    }
+
+                    // Extract provided properties into the composition
+                    repository.ExtractComposition(contentType, compositionType, propertyAliases);
+                }
+
+                UpdateContentXmlStructure(contentType);
+            }
+
+            Audit(AuditType.Custom, "Extract content type composition performed by user", userId, -1);
+            return compositionType;
+        }
+
+        /// <summary>
+        /// Extracts a composition from a content tye
+        /// </summary>
+        /// <param name="mediaType"><see cref="IMediaType"/> to extract composition from</param>
+        /// <param name="propertyAliases">Aliases of properties to move to composition type</param>
+        /// <param name="extractIntoType"><see cref="IMediaType"/> to extract composition into</param>
+        /// <param name="nameOfNewType">Name of new composition type if creating a new one</param>
+        /// <param name="userId">Optional Id of the User deleting the ContentType</param>
+        /// <returns>The created composition type</returns>
+        public IMediaType ExtractComposition(IMediaType mediaType, string[] propertyAliases,
+            IMediaType extractIntoType = null, string nameOfNewType = null, int userId = 0)
+        {
+            Mandate.ParameterNotNull(mediaType, "mediaType");
             Mandate.ParameterNotNullOrEmpty(propertyAliases, "propertyAliases");
+            MandateExtractCompositionParametersValid(mediaType, propertyAliases, extractIntoType, nameOfNewType);
+
+            IMediaType compositionType;
+            using (new WriteLock(Locker))
+            {
+                var uow = UowProvider.GetUnitOfWork();
+                using (var repository = RepositoryFactory.CreateMediaTypeRepository(uow))
+                {
+                    if (extractIntoType == null)
+                    {
+                        // Create new composition type
+                        compositionType = new MediaType(mediaType.ParentId)
+                        {
+                            Alias = nameOfNewType.ToSafeAlias(true),
+                            Name = nameOfNewType,
+                            Icon = "icon-document",
+                        };
+                        repository.AddOrUpdate(compositionType);
+                        uow.Commit();
+                    }
+                    else
+                    {
+                        compositionType = extractIntoType;
+                    }
+
+                    // Extract provided properties into the composition
+                    repository.ExtractComposition(mediaType, compositionType, propertyAliases);
+                }
+
+                UpdateContentXmlStructure(mediaType);
+            }
+
+            Audit(AuditType.Custom, "Extract media type composition performed by user", userId, -1);
+            return compositionType;
+        }
+
+        /// <summary>
+        /// Helper to provide common validation of the parameters passed to extract composition methods
+        /// </summary>
+        /// <param name="contentType"><see cref="IContentTypeComposition"/> to extract composition from</param>
+        /// <param name="propertyAliases">Aliases of properties to move to composition type</param>
+        /// <param name="extractIntoType"><see cref="IContentTypeComposition"/> to extract composition into</param>
+        /// <param name="nameOfNewType">Name of new composition type if creating a new one</param>
+        private static void MandateExtractCompositionParametersValid(IContentTypeComposition contentType, string[] propertyAliases, 
+            IContentTypeComposition extractIntoType, string nameOfNewType)
+        {
+            Mandate.ParameterNotNullOrEmpty(propertyAliases, "propertyAliases");
+
             if (extractIntoType == null && string.IsNullOrEmpty(nameOfNewType))
             {
                 throw new InvalidOperationException(
@@ -923,110 +1020,6 @@ namespace Umbraco.Core.Services
                         "The selected type to extract a composition into contains properties with aliases matching those of the current type. " +
                         "In that situation those properties must be selected as part of the extract composition operation.");
                 }
-            }
-
-            IContentType compositionType;
-            using (new WriteLock(Locker))
-            {
-                var uow = UowProvider.GetUnitOfWork();
-                if (extractIntoType == null)
-                {
-                    using (var repository = RepositoryFactory.CreateContentTypeRepository(uow))
-                    {
-                        // Create new composition type
-                        compositionType = new ContentType(contentType.ParentId)
-                        {
-                            Alias = nameOfNewType.ToSafeAlias(true),
-                            Name = nameOfNewType,
-                            Icon = "icon-document",
-                        };
-                        repository.AddOrUpdate(compositionType);
-                        uow.Commit();
-                    }
-                }
-                else
-                {
-                    compositionType = extractIntoType;
-                }
-
-                // Extract provided properties into the composition
-                ExtractPropertiesToComposition(uow, contentType, compositionType, propertyAliases);
-
-                UpdateContentXmlStructure(contentType);
-            }
-
-            Audit(AuditType.Custom, "Extract content type composition performed by user", userId, -1);
-            return compositionType;
-        }
-
-        /// <summary>
-        /// Extracts a composition from a content tye
-        /// </summary>
-        /// <param name="mediaType"><see cref="IMediaType"/> to extract composition from</param>
-        /// <param name="propertyAliases">Aliases of properties to move to composition type</param>
-        /// <param name="extractIntoType"><see cref="IMediaType"/> to extract composition into</param>
-        /// <param name="nameOfNewType">Name of new composition type if creating a new one</param>
-        /// <param name="userId">Optional Id of the User deleting the ContentType</param>
-        /// <returns>The created composition type</returns>
-        public IMediaType ExtractComposition(IMediaType mediaType, string[] propertyAliases,
-            IMediaType extractIntoType = null, string nameOfNewType = null, int userId = 0)
-        {
-            Mandate.ParameterNotNull(mediaType, "mediaType");
-            Mandate.ParameterNotNullOrEmpty(propertyAliases, "propertyAliases");
-            if (extractIntoType == null && string.IsNullOrEmpty(nameOfNewType))
-            {
-                throw new InvalidOperationException(
-                    "When extracting a media type either the type to extract into or the name for a new one must be provided.");
-            }
-
-            IMediaType compositionType;
-            using (new WriteLock(Locker))
-            {
-                var uow = UowProvider.GetUnitOfWork();
-                if (extractIntoType == null)
-                {
-                    using (var repository = RepositoryFactory.CreateMediaTypeRepository(uow))
-                    {
-                        // Create new composition type
-                        compositionType = new MediaType(mediaType.ParentId)
-                        {
-                            Alias = nameOfNewType.ToSafeAlias(true),
-                            Name = nameOfNewType,
-                            Icon = "icon-document",
-                        };
-                        repository.AddOrUpdate(compositionType);
-                        uow.Commit();
-                    }
-                }
-                else
-                {
-                    compositionType = extractIntoType;
-                }
-
-                // Extract provided properties into the composition
-                ExtractPropertiesToComposition(uow, mediaType, compositionType, propertyAliases);
-
-                UpdateContentXmlStructure(mediaType);
-            }
-
-            Audit(AuditType.Custom, "Extract media type composition performed by user", userId, -1);
-            return compositionType;
-        }
-
-        /// <summary>
-        /// Helper to extract the composition properties for content and media types
-        /// </summary>
-        /// <param name="uow">Unit of work instance</param>
-        /// <param name="contentType"><see cref="IContentTypeComposition"/> to extract composition from</param>
-        /// <param name="compositionType">Created type for the composition</param>
-        /// <param name="propertyAliases">Aliases of properties to move to composition type</param>
-        private void ExtractPropertiesToComposition(IDatabaseUnitOfWork uow, 
-            IContentTypeComposition contentType, IContentTypeComposition compositionType, 
-            string[] propertyAliases)
-        {
-            using (var repository = RepositoryFactory.CreateContentTypeRepository(uow))
-            {
-                repository.ExtractComposition(contentType, compositionType, propertyAliases);
             }
         }
